+++ NEWS.md
@@ -3,6 +3,7 @@
 - Convert marc field 545 to Resource description [MODLD-337](https://folio-org.atlassian.net/browse/MODLD-337)
 - Derive MARC 300 from Graph [MODLD-401](https://folio-org.atlassian.net/browse/MODLD-401)
 - Convert marc field 586 (Awards Note) to Resource description [MODLD-366](https://folio-org.atlassian.net/browse/MODLD-366)
+- Create Place resource from marc auth field 151 when $v, $x, $y, $z not present [MODLD-529](https://folio-org.atlassian.net/browse/MODLD-529)
 
 ## 1.0.2 (03-12-2025)
 - Initial release